Move event listeners to ``sphinx.ext.autodoc._event_listeners``
